refactor : move `MySpeciesAutoComplete`
